requirements.txt
praw
psycopg2
python-dotenv